Additional refactoring
Some renames and deduplication
